refact: cell 에 전달하는 바인딩값을 struct 으로 묶어서 전달 (#9)
